delete account api integration

+++ client/src/components/modals/DeleteAccountModal.tsx
@@ -1,23 +1,33 @@
 import ModalOverlay from "./Modal";
 import { useAppDispatch, useAppSelector } from "../../../hooks/reduxHooks";
 import { setIsDeleteAccount } from "../../../redux/slices/authSlices";
+import { deleteAccountAction } from "../../../redux/actions/authActions";
+import { useNavigate } from "react-router-dom";
 
 const DeleteAccountModal = () => {
-  const { isDeleteAccount } = useAppSelector((state) => state.auth);
+  const { isDeleteAccount, deleteAccountLoading } = useAppSelector(
+    (state) => state.auth
+  );
   const dispatch = useAppDispatch();
+  const navigate = useNavigate();
   if (!isDeleteAccount) return null;
 
   const closeModal = () => {
+    if (deleteAccountLoading) return;
     dispatch(setIsDeleteAccount(false));
   };
 
   const deleteAccountHandler = () => {
-    // TODO: Delete Account
+    if (deleteAccountLoading) return;
+    dispatch(deleteAccountAction(navigate));
   };
-
   return (
     <ModalOverlay onClose={closeModal}>
-      <div className={`flex flex-col gap-4`}>
+      <div
+        className={`flex flex-col gap-4 opacity-100 ${
+          deleteAccountLoading ? "opacity-65" : "opacity-100"
+        }`}
+      >
         <h1 className="text-red-800 font-bold text-2xl">
           Deleting Your Account!
         </h1>
@@ -35,13 +45,21 @@ const DeleteAccountModal = () => {
           <button
             onClick={closeModal}
             style={{ backgroundColor: "var(--main-color)" }}
-            className={`bg-red-200 text-gray-700 py-2 px-4 rounded-md font-semibold outline-none`}
+            className={`bg-red-200 text-gray-700 py-2 px-4 rounded-md font-semibold outline-none opacity-100 ${
+              deleteAccountLoading
+                ? "opacity-65 cursor-default"
+                : "opacity-100 cursor-pointer"
+            }`}
           >
             Cancel
           </button>
           <button
             onClick={deleteAccountHandler}
-            className={` bg-red-700 text-white py-2 px-4 rounded-md font-semibold outline-none`}
+            className={` bg-red-700 text-white py-2 px-4 rounded-md font-semibold outline-none opacity-100 ${
+              deleteAccountLoading
+                ? "opacity-65 cursor-default"
+                : "opacity-100 cursor-pointer"
+            }`}
           >
             Delete Account
           </button>
